requirements.txt
torch
pillow
numpy
